Created CatModule service to handle downloads and stuff
